% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_diversity.R
\name{calc_TDD_THD}
\alias{calc_TDD_THD}
\title{Calculate the Shannon-Wiener Index for a Continuous Variable}
\usage{
calc_TDD_THD(data, variable, class_width, log_base = exp(1), normalize = TRUE)
}
\arguments{
\item{data}{A data frame containing the data to analyze.}

\item{variable}{The column name (as a string) containing the continuous variable to analyze.}

\item{class_width}{The width of the classes for discretizing the continuous variable.}

\item{log_base}{The base of the logarithm used in the calculation of the Shannon-Wiener index. The default is natural logarithm (e = exp(1)).}

\item{normalize}{Logical value indicating whether to normalize the index (default is TRUE). If TRUE, the index will be normalized by dividing by the natural logarithm of the number of classes.}
}
\value{
The Shannon-Wiener index for the discretized variable, as a numeric value.
}
\description{
This function calculates the Shannon-Wiener index (a measure of diversity) for a continuous variable.
The continuous variable is discretized into classes based on the specified \code{class_width}, and the Shannon-Wiener index
is computed for the resulting frequency distribution. The Shannon-Wiener index is commonly used in ecology and
environmental studies to assess the diversity of a population or ecosystem.
}
\details{
The Shannon-Wiener index is calculated using the formula:
\deqn{H' = - \sum p_i \log_b(p_i)},
where \emph{p_i} is the proportion of observations in the ith class, and \emph{b} is the logarithmic base.
The index is normalized by dividing by \(\log_b(C)\), where \(C\) is the number of classes.
}
\examples{
# Example usage
data <- data.frame(variable = c(1.5, 2.3, 3.1, 1.8, 2.5, 3.3))
calc_TDD_THD(data, variable = "variable", class_width = 0.5)

}
