% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_diversity.R
\name{calculate_SDI}
\alias{calculate_SDI}
\title{Stand Density Index (SDI) Calculation}
\usage{
calculate_SDI(N, G, b = 1.605, type = "coetanei")
}
\arguments{
\item{N}{A numeric value representing tree density (trees per hectare).}

\item{G}{A numeric value representing the total basal area (m² per hectare), used to compute QMD.}

\item{b}{A numeric value for the scaling exponent (default is 1.605). This parameter adjusts the relationship
between tree density and diameter.}

\item{type}{A character string specifying the type of forest stand. Options are:
\item "coetanei" for even-aged stands (default),
\item "disetanei" for uneven-aged stands.}
}
\value{
A numeric value representing the calculated Stand Density Index (SDI) for the given forest stand.
}
\description{
This function calculates the Stand Density Index (SDI) for a forest plot based on tree density
and quadratic mean diameter (QMD). The SDI is commonly used to assess stand density in forest
management, and it can be calculated for different forest types, such as even-aged (coetanei) or
uneven-aged (disetanei) stands.
}
\details{
The formula for SDI is:
\deqn{SDI = (N * (QMD^b))}, where:
\itemize{
\item N is the tree density (trees per hectare),
\item QMD is the quadratic mean diameter (cm),
\item b is a scaling exponent, typically set to 1.605 (default).
}

The SDI provides an estimate of the stand's density, which helps evaluate its competition and growth potential.
}
\examples{
N <- 500
G <- 40
calculate_SDI(N, G)  # for even-aged stand

N2 <- 300
G2 <- 30
calculate_SDI(N2, G2, type = "disetanei")  # for uneven-aged stand

}
