% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_diversity.R
\name{CI_plot_1000}
\alias{CI_plot_1000}
\title{Calculate the Structural Diversity Index (CI) for a given dataset of trees.}
\usage{
CI_plot_1000(data, th_col, ba_col, specie_col, plot_area)
}
\arguments{
\item{data}{A data frame containing tree measurements.}

\item{th_col}{A string representing the name of the column for tree heights.}

\item{ba_col}{A string representing the name of the column for basal area.}

\item{specie_col}{A string representing the name of the column for tree species.}

\item{plot_area}{A numeric value representing the area of the plot in square meters.}
}
\value{
A numeric value representing the Structural Diversity Index (CI).
}
\description{
This function computes the CI based on tree height, basal area, and species composition
from a specified plot area. It is useful for ecologists and forest managers assessing
the diversity and structural complexity of forest stands.
}
\examples{
# Example usage:
data <- as.data.frame(list(th = c(15, 20, 25, 10, 30),
                          ba = c(5, 10, 15, 5, 20),
                          specie = c("fir", "european beech", "fir", "european beech", "european beech")))

results<- CI_plot_1000(data, "th", "ba", "specie", plot_area = 100)

print(results)

}
