% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_structure.R
\name{DominantTreeByPlot}
\alias{DominantTreeByPlot}
\title{Dominant Tree Species Detection in Forest Stands Based on Basal Area}
\usage{
DominantTreeByPlot(data, BA_col, specie_col)
}
\arguments{
\item{data}{A data frame containing tree data with a column for species and basal area.}

\item{BA_col}{The name of the column in the data frame representing basal area (e.g., "Sum_BA_m2").}

\item{specie_col}{The name of the column in the data frame representing tree species (e.g., "specie").}
}
\value{
A character vector containing:
\item{Pure_Mixed_Stands}{The type of stand: "Pure" if a single species dominates with more than 50% of the total
basal area, "Mixed" otherwise.}
\item{DomTreeSpecies}{The name of the dominant species if the stand is "Pure", or "other broadleaf/coniferous"
if the stand is "Mixed".}
}
\description{
This function identifies the dominant tree species in a forest stand and classifies the stand as either "Pure" or
"Mixed" based on basal area accumulation. The classification of the stand as "Pure" is based on the condition that
one species contributes more than 50% of the total basal area in the stand. If no species exceeds this threshold,
the stand is classified as "Mixed". The dominant species is identified as the species with the highest basal area
percentage if the stand is "Pure". For "Mixed" stands, the dominant species is classified as "other broadleaf/coniferous".
}
\details{
To classify the plot as "Pure" or "Mixed", the function calculates the total basal area of the stand and
determines the percentage of basal area contributed by each species. If any species exceeds 50% of the total basal area,
it is considered the dominant species, and the stand is classified as "Pure". In cases where no species exceeds
50%, the stand is classified as "Mixed". For mixed stands, the dominant tree species is labeled as "other broadleaf/coniferous"
to represent the mixed composition of the stand.
}
\examples{
data <- data.frame(specie_col = c("Pine", "Oak", "Birch"), BA_col = c(30, 20, 5))
results <- DominantTreeByPlot(data=data, BA_col="BA_col", specie_col="specie_col")
print(results)

}
