% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_structure.R
\name{DomTreeSpeciesDetection}
\alias{DomTreeSpeciesDetection}
\title{Detect Dominant Tree Species}
\usage{
DomTreeSpeciesDetection(data)
}
\arguments{
\item{data}{A data frame containing tree data with a column for species and basal area.}
}
\value{
A character vector containing the type of stand ("Pure" or "Mixed") and the dominant species or "Mixed" if no species exceeds 50\% of the basal area.
}
\description{
This function detects the dominant tree species in a stand based on basal area. If more than 50\% of the
stand's basal area is made up of one species, it is considered the dominant species.
}
\examples{
data <- data.frame(specie = c("Pine", "Oak", "Birch"),Sum_BA_m2 = c(30, 20, 5))
DomTreeSpeciesDetection(data)
}
