% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{load_data}
\alias{load_data}
\title{Load_data
 Helper Functions for Data Processing in Shiny App}
\usage{
load_data(file_path)
}
\arguments{
\item{file_path}{The file path to the dataset (Excel format expected).}

\item{data}{A dataframe containing the data to be processed.}

\item{columns}{A character vector specifying the columns to retain in the subset.}

\item{new_names}{A character vector with new column names (must match the number of columns in `data`).}
}
\value{
The functions return processed data:

- `load_data()`: Returns a dataframe loaded from an Excel file.
- `subset_data()`: Returns a dataframe with only the specified columns.
- `rename_columns()`: Returns a dataframe with renamed columns.
}
\description{
A set of functions to load, subset, and rename columns in datasets, designed for use in a Shiny application.
}
\examples{
# Load data
data <- load_data("C:/Users/calvi/OneDrive/Desktop/App_Forwards/app/Data/Example.xlsx")

# Subset specific columns
subset <- subset_data(data, c("TreeDiameter", "TreeHeight"))

# Rename columns
renamed_data <- rename_columns(subset, c("NewName1", "NewName2"))

}
