% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{process_all_deadwood}
\alias{process_all_deadwood}
\title{process_all_deadwood
Process All Deadwood Data}
\usage{
process_all_deadwood(data, input)
}
\arguments{
\item{data}{A `data.frame` containing forest inventory data.}

\item{input}{A list containing column names and parameters required for processing.}
}
\value{
A processed dataset with estimated deadwood values.
}
\description{
Processes deadwood data by subsetting relevant columns and applying the deadwood estimation model.
}
\examples{
data <- data.frame(
  ForManInt = c(1, 2, 1),
  plot = c("A", "B", "C"),
  L_tot_col = c(10, 15, 20),
  Dmax_col = c(0.5, 0.6, 0.7),
  Dmin_col = c(0.3, 0.4, 0.5)
)

input <- list(
  I5_ForManInt_col = "ForManInt",
  I5_plot_col = "plot",
  I5_L_col = "L_tot_col",
  I5_Dmax_col = "Dmax_col",
  I5_Dmin_col = "Dmin_col",
  I5_Provide_ForManInt = "some_value",
  I5_Provide_plot = "some_value",
  I5_plot_size_col = "some_value"
)

results <- process_all_deadwood(data, input)

}
