% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{process_carbon_stock}
\alias{process_carbon_stock}
\title{process_carbon_stock
 Process Carbon Stock Model}
\usage{
process_carbon_stock(data, input)
}
\arguments{
\item{data}{A `data.frame` containing forest inventory data.}

\item{input}{A list containing column names and parameters required for processing.}
}
\value{
A processed dataset with estimated carbon stock values.
}
\description{
Processes carbon stock data by subsetting relevant columns and applying the carbon stock calculator.
}
\examples{
data <- data.frame(
  ForManInt = c(1, 2, 1),
  plot = c("A", "B", "C"),
  dom_col = c(10, 15, 20),
  vol_m3_ha = c(200, 300, 400)
)

input <- list(
  I2_ForManInt_col = "ForManInt",
  I2_plot_col = "plot",
  I2_dom_col = "dom_col",
  I2_vol_col = "vol_m3_ha",
  I2_Provide_ForManInt = "some_value",
  I2_Provide_plot = "some_value"
)

results <- process_carbon_stock(data, input)

}
