% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{process_growing_stock}
\alias{process_growing_stock}
\title{process_growing_stock
Process Growing Stock Model}
\usage{
process_growing_stock(data, input)
}
\arguments{
\item{data}{A `data.frame` containing forest inventory data.}

\item{input}{A list containing column names and parameters required for processing.}
}
\value{
A processed dataset with calculated stem volume estimates.
}
\description{
Processes growing stock data by subsetting relevant columns and applying the stem volume calculator.
}
\examples{
data <- data.frame(
  ForManInt = c(1, 2, 1),
  plot = c("A", "B", "C"),
  th = c(15, 20, 25),
  dbh = c(30, 40, 50),
  specie = c("Fagus", "Quercus", "Pinus")
)

input <- list(
  I1_ForManInt_col = "ForManInt",
  I1_plot_col = "plot",
  I1_height_col = "th",
  I1_dbh_col = "dbh",
  I1_specie_col = "specie",
  I1_Provide_ForManInt = "some_value",
  I1_Provide_plot = "some_value",
  I1_plot_size_col = 500
)

results <- process_growing_stock(data, input)

}
