% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{process_lying_deadwood}
\alias{process_lying_deadwood}
\title{process_lying_deadwood
 Process Lying Deadwood Data}
\usage{
process_lying_deadwood(data, input)
}
\arguments{
\item{data}{A `data.frame` containing forest inventory data.}

\item{input}{A list containing column names and parameters required for processing.}
}
\value{
A processed dataset with estimated lying deadwood values.
}
\description{
Processes lying deadwood data by subsetting relevant columns and applying the deadwood estimation model.
}
\examples{
data <- data.frame(
  ForManInt = c(1, 2, 1),
  plot = c("A", "B", "C"),
  L_tot_col = c(10, 15, 20),
  Dhalf_col = c(0.3, 0.4, 0.5),
  TH_tot_col = c(1.5, 2.0, 2.5),
  DBH_col = c(25, 30, 35)
)

input <- list(
  I3_ForManInt_col = "ForManInt",
  I3_plot_col = "plot",
  I3_L_tot_col = "L_tot_col",
  I3_Dhalf_col = "Dhalf_col",
  I3_TH_tot_col = "TH_tot_col",
  I3_DBH_col = "DBH_col",
  I3_Provide_ForManInt = "some_value",
  I3_Provide_plot = "some_value",
  I3_LDT_CWD_option = "some_option",
  I3_plot_size_col = "some_value"
)

results <- process_lying_deadwood(data, input)

}
