% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{process_standing_deadwood}
\alias{process_standing_deadwood}
\title{process_standing_deadwood
Process Standing Deadwood Data}
\usage{
process_standing_deadwood(data, input)
}
\arguments{
\item{data}{A `data.frame` containing forest inventory data.}

\item{input}{A list containing column names and parameters required for processing.}
}
\value{
A processed dataset with estimated standing deadwood values.
}
\description{
Processes standing deadwood data by subsetting relevant columns and applying the deadwood estimation model.
}
\examples{
data <- data.frame(
  ForManInt = c(1, 2, 1),
  plot = c("A", "B", "C"),
  L_tot_col = c(10, 15, 20),
  Dhalf_col = c(0.3, 0.4, 0.5),
  TH_tot_col = c(1.5, 2.0, 2.5),
  DBH_col = c(25, 30, 35)
)

input <- list(
  I4_ForManInt_col = "ForManInt",
  I4_plot_col = "plot",
  I4_L_tot_col = "L_tot_col",
  I4_Dhalf_col = "Dhalf_col",
  I4_TH_tot_col = "TH_tot_col",
  I4_DBH_col = "DBH_col",
  I4_Provide_ForManInt = "some_value",
  I4_Provide_plot = "some_value",
  I4_SDT_SNAG_option = "some_option",
  I4_plot_size_col = "some_value"
)

results <- process_standing_deadwood(data, input)

}
