% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{rename_columns}
\alias{rename_columns}
\title{rename_columns
third helper Functions for Data Processing in Shiny App}
\usage{
rename_columns(data, new_names)
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}

\item{new_names}{A character vector with new column names (must match the number of columns in `data`).}

\item{file_path}{The file path to the dataset (Excel format expected).}

\item{columns}{A character vector specifying the columns to retain in the subset.}
}
\value{
The functions return processed data:

- `rename_columns()`: Returns a dataframe with renamed columns.
}
\description{
A set of functions to load, subset, and rename columns in datasets, designed for use in a Shiny application.
}
