% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standing_lying_deadwood.R
\name{Apply_LyingDeadwood}
\alias{Apply_LyingDeadwood}
\title{Apply Lying Deadwood or Coarse Woody Debris Volume Calculation}
\usage{
Apply_LyingDeadwood(
  data,
  ForManInt_option,
  ForManInt,
  plot_option,
  plot_col,
  LDT_CWD_option,
  L_tot_col,
  Dhalf_col,
  TH_tot_col,
  DBH_col,
  plot_area
)
}
\arguments{
\item{data}{A data frame containing tree attributes for each plot. The data must include
relevant columns such as DBH, tree height, and forest management intervention information.
Additional parameters for the volume calculations, such as L_tot and Dhalf, are also required for CWD.}

\item{ForManInt_option}{A character string ("Yes" or "No") indicating whether forest management intervention
data is available in the dataset. Defaults to "No" if not provided. If "Yes", the column specified in
the \code{ForManInt} argument will be used to group the results.}

\item{ForManInt}{A character string representing the name of the column in the dataset that specifies the
forest management intervention (e.g., thinning, clearcut). Used if \code{ForManInt_option} is "Yes".}

\item{plot_option}{A character string ("Yes" or "No") indicating whether plot-level information is available.
Defaults to "No" if not provided. If "Yes", the data will be grouped by plots and summarized accordingly.}

\item{plot_col}{A character string representing the name of the column in the dataset that specifies the plot identifier.
Used if \code{plot_option} is "Yes".}

\item{LDT_CWD_option}{A character string specifying which type of deadwood to calculate ("ldt" for Lying Deadwood or
"cwd" for Coarse Woody Debris). The corresponding calculation will be applied based on this option.}

\item{L_tot_col}{A character string representing the name of the column in the dataset for the total tree length
(L_tot), used in the CWD volume calculation.}

\item{Dhalf_col}{A character string representing the name of the column in the dataset for the half-diameter
(Dhalf), used in the CWD volume calculation.}

\item{TH_tot_col}{A character string representing the name of the column in the dataset for the total tree height
(TH_tot), used in the LDT volume calculation.}

\item{DBH_col}{A character string representing the name of the column in the dataset for the diameter at breast height
(DBH), used in the LDT volume calculation.}

\item{plot_area}{A numeric value representing the area of the plot in square meters. This is used to calculate
the volume per hectare.}
}
\value{
A list containing two data frames:
\item{output_plot}{A data frame with the calculated volume for each plot, including the volume of LDT or CWD per hectare.}
\item{output_ForManInt}{A summary data frame showing the average volume of LDT or CWD per hectare, grouped by forest management intervention.}
}
\description{
This function calculates the volumes of Lying Deadwood (LDT) or Coarse Woody Debris (CWD)
based on tree attributes such as diameter at breast height (DBH), tree height, and
additional parameters like the total length of the tree (L_tot) and its half-diameter
(Dhalf). The calculation is performed for each plot, and the function can group the results
by forest management interventions.
}
\details{
This function allows for the calculation of volumes of Lying Deadwood (LDT) or Coarse Woody Debris (CWD)
based on tree attributes, and can group results by plot or forest management intervention type. The specific
calculation for LDT or CWD depends on the user-specified option (\code{LDT_CWD_option}).
The volume per hectare is calculated by considering the area of the plot.
}
\examples{
\dontrun{
data <- read.csv("forest_inventory.csv")
result <- Apply_LyingDeadwood(data, "Yes", "Thinning", "Yes", "plot_id", "ldt", "L_tot", "Dhalf", "TH_tot", "DBH", 10000)
}

}
