% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_diversity.R
\name{ForStrSpecDiv}
\alias{ForStrSpecDiv}
\title{Forest Structure and Species Diversity Indices}
\usage{
ForStrSpecDiv(data, dbh_col, th_col, ba_col, specie_col, plot_area)
}
\arguments{
\item{data}{A data frame containing the forest inventory data. It should include columns for
diameter at breast height (dbh), tree height (th), basal area (ba), and species.}

\item{dbh_col}{A character string specifying the column name for diameter at breast height (dbh)
in centimeters.}

\item{th_col}{A character string specifying the column name for tree height (th) in meters.}

\item{ba_col}{A character string specifying the column name for basal area (ba) in square meters.}

\item{specie_col}{A character string specifying the column name for species names.}

\item{plot_area}{A numeric value indicating the area of the forest plot in square meters (m²).}
}
\value{
A data frame containing the calculated diversity and structural indices for the forest plot.
The returned data frame includes:
\itemize{
\item Simpson's diversity index (for dbh, th, and species richness)
\item Shannon's diversity index (for dbh, th, and species richness)
\item Gini index for basal area and diameter
\item Coefficient of variation for basal area
\item Vertical Evenness Index (VEm)
\item Stand Density Index (SDI)
}
}
\description{
This function calculates various forest diversity indices and tree structural metrics,
including Simpson and Shannon diversity indices for diameter at breast height (dbh),
tree height (th), and species richness. It also computes the Gini index for basal area and
diameter, the coefficient of variation for basal area, and the Vertical Evenness Index (VEm).
Additionally, the function calculates the Stand Density Index (SDI).
}
\details{
This function performs multiple calculations on a given forest inventory dataset to
assess forest structural complexity and species diversity. The calculated indices help in
understanding the spatial and vertical distribution of trees and their structural variability.
These metrics are useful for forest management, ecological assessments, and biodiversity studies.
}
\examples{
# Example usage:
data <- as.data.frame(list(
  dbh_col = c(10, 15, 18, 30, 35, 45),     # Diameter at breast height (DBH) in cm
  th_col = c(15, 18, 19, 25, 26, 32),      # Tree height in meters
  ba_col = c(0.007854, 0.017671, 0.025446, 0.070685, 0.096211, 0.159154),  # Basal Area (calculated)
  specie_col = c("norway spruce", "norway spruce", "norway spruce", "norway spruce", "norway spruce", "norway spruce")
))
plot_area = 530
results <- ForStrSpecDiv(data = data,
                        dbh_col = "dbh_col",
                        th_col = "th_col",
                        ba_col = "ba_col",
                        specie_col = "specie_col",
                        plot_area = plot_area)

print(results)

}
